Add files via upload
I have managed to produce a working version of SDM for ArcGIS 10.x based on an update released back in November. I’ve tested it successfully on 10.3 and 10.5. As far as I can tell, I've resolved most of the previously reported issues except for the hard coded limitations for logistic regression in sdmlr.exe and I also didn't touch neural networks. I don't think it will work on Pro, because Pro uses a different version of Python, but to the best of my knowledge it should work on any 10.x.

I am attaching a copy of the toolbox that has resolved all the issues we were having for Calculate Weights, Calculate Response, Grand WofE, and as much of Logistic Regression as is possible. The xml files for the tools have been updated where necessary (explicitly noting the limitations for LR). 

Specific fixes include:

* Perfect correlation is dealt with for all cases
* Having pattern area less than unit cell size now has a more appropriate patch
* Fixed up the generalized weights in the weights tables for the "out" class in perfect correlation cases
* Grand WofE and Logistic Regression now explicitly allow you to specify ascending or descending, not just "ordered".
* Fixed it so that it takes a raster mask